+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.6.0] - 2023-12-24
+
+### Added
+
+- Add support for globs in `--rules` argument
+
+### Fixed
+
+- Error preventing usage of `--dry-run` and `--dry-run-short`
+
 ## [0.5.0] - 2023-12-22
 
 ### Added